+++ app/src/main/java/com/mytech/salesvisit/util/VectorTOJsonConverter.java
@@ -6,44 +6,38 @@ import com.google.gson.JsonObject;
 import java.util.Vector;
 
 public class VectorTOJsonConverter {
-    public JsonArray getJsonArray(Vector[] v)
-    {
-        JsonArray jsonArray=new JsonArray();
+    public JsonArray getJsonArray(Vector[] v) {
+        JsonArray jsonArray = new JsonArray();
         JsonObject jsonObject;
-        try{
-            for(int i=0;i<v.length;i++)
-            {
-                jsonObject=new JsonObject();
-                jsonObject.addProperty("ItemId", ""+v[i].get(9));
-                    jsonObject.addProperty("OrderId",  0);
-                    jsonObject.addProperty("SrNo",  1);
-                    jsonObject.addProperty("ProductId",  Integer.parseInt(v[i].get(1).toString()));
-                    jsonObject.addProperty("Specification",  v[i].get(6).toString());
-                    jsonObject.addProperty("Qty",  Integer.parseInt(v[i].get(2).toString()));
-                    jsonObject.addProperty("UOMId",  Integer.parseInt(v[i].get(3).toString()));
-                    jsonObject.addProperty("Rate",  Double.parseDouble(v[i].get(4).toString()));
-                    jsonObject.addProperty("StatusId",  1);
-                    jsonObject.addProperty("IsRemoved",  false);
-                    jsonObject.addProperty("OtherProduct",  v[i].get(7).toString());
+        try {
+            for (int i = 0; i < v.length; i++) {
+                jsonObject = new JsonObject();
+                jsonObject.addProperty("ItemId", "" + v[i].get(9));
+                jsonObject.addProperty("OrderId", 0);
+                jsonObject.addProperty("SrNo", 1);
+                jsonObject.addProperty("ProductId", Integer.parseInt(v[i].get(1).toString()));
+                jsonObject.addProperty("Specification", v[i].get(6).toString());
+                jsonObject.addProperty("Qty", Float.parseFloat(v[i].get(2).toString()));
+                jsonObject.addProperty("UOMId", Integer.parseInt(v[i].get(3).toString()));
+                jsonObject.addProperty("Rate", Double.parseDouble(v[i].get(4).toString()));
+                jsonObject.addProperty("StatusId", 1);
+                jsonObject.addProperty("IsRemoved", false);
+                jsonObject.addProperty("OtherProduct", v[i].get(7).toString());
                 jsonArray.add(jsonObject);
             }
-return jsonArray;
-        }catch(Exception e)
-        {
+            return jsonArray;
+        } catch (Exception e) {
             return null;
         }
     }
 
 
-
-    public JsonArray getJsonArrayTerms(Vector[] v)
-    {
-        JsonArray jsonArray=new JsonArray();
+    public JsonArray getJsonArrayTerms(Vector[] v) {
+        JsonArray jsonArray = new JsonArray();
         JsonObject jsonObject;
-        try{
-            for(int i=0;i<v.length;i++)
-            {
-                jsonObject=new JsonObject();
+        try {
+            for (int i = 0; i < v.length; i++) {
+                jsonObject = new JsonObject();
 
                 /*
 
@@ -60,17 +54,16 @@ return jsonArray;
                  */
 
                 jsonObject.addProperty("TermId", Integer.parseInt(v[i].get(0).toString().trim()));
-                jsonObject.addProperty("SrNo",  Integer.parseInt(v[i].get(1).toString().trim()));
-                jsonObject.addProperty("OrderId",  Integer.parseInt(v[i].get(2).toString().trim()));
-                jsonObject.addProperty("ParticularId",  Integer.parseInt(v[i].get(3).toString()));
-                jsonObject.addProperty("Condition",  v[i].get(4).toString());
-                jsonObject.addProperty("IsRemoved",  Boolean.parseBoolean(v[i].get(5).toString()));
+                jsonObject.addProperty("SrNo", Integer.parseInt(v[i].get(1).toString().trim()));
+                jsonObject.addProperty("OrderId", Integer.parseInt(v[i].get(2).toString().trim()));
+                jsonObject.addProperty("ParticularId", Integer.parseInt(v[i].get(3).toString()));
+                jsonObject.addProperty("Condition", v[i].get(4).toString());
+                jsonObject.addProperty("IsRemoved", Boolean.parseBoolean(v[i].get(5).toString()));
 
                 jsonArray.add(jsonObject);
             }
             return jsonArray;
-        }catch(Exception e)
-        {
+        } catch (Exception e) {
             return null;
         }
     }
